Small changes to build arm64 simulator library
Some small changes to the build scripts and configs to make it so that
we can build arm64 simulator binaries. No big refactorings, just making
it work well enough to build.

This is the basis of release 0.8.1

Note that because of some... interesting behaviors of cocoapods, we
needed to check the artifacts into the repo rather than rely on a
release. tl;dr; if you're using a git source, it ignores the source in
the podspec, and some tools don't like it if you have an https:// in the
`:podspec` attribute of the Podfile.